Share workflow job execution selection
