+++ libs/client/workflows/src/core/entities/job.ts
@@ -101,6 +101,30 @@ export function isTerminalJobStatus(status: JobStatus): boolean {
   return TERMINAL_JOB_STATUS_SET.has(status);
 }
 
+export function resolveJobExecution(
+  job: Job,
+  jobExecutionId: string | undefined,
+): JobExecution | undefined {
+  const selectedExecution = jobExecutionId
+    ? job.jobExecutions.find((jobExecution) => jobExecution.id === jobExecutionId)
+    : undefined;
+  if (selectedExecution) return selectedExecution;
+
+  return defaultJobExecution(job);
+}
+
+export function defaultJobExecution(job: Job): JobExecution | undefined {
+  const runningExecution = job.jobExecutions.find(
+    (jobExecution) => jobExecution.status === 'running',
+  );
+  if (runningExecution) return runningExecution;
+
+  return job.jobExecutions.reduce<JobExecution | undefined>((latest, jobExecution) => {
+    if (!latest) return jobExecution;
+    return jobExecution.sequence > latest.sequence ? jobExecution : latest;
+  }, undefined);
+}
+
 export function toJob(dto: WorkflowRunJobDetailDto): Job {
   const jobExecutions = dto.job_executions.map(toJobExecution);
